Introduce a new pool property 'dedupentrymax' to limit entries in the DDT
You can also limit by size, using 'dedupsizemax', which will calculate
approximately the correct number of entries, and use that value.

The main reason to track entries is the size of the DDT does not shrink
when entries are removed, so tracking entries works much more reliably.

Track the current size of the DDT, and the average entry size in spa_t

Once the maximum number of entries is reached, new entries will not be
deduped. Existing entries are still updated and continue to dedup.

Signed-off-by: Sean Fagan <sean.fagan@klarasystems.com>
Signed-off-by: Allan Jude <allan@klarasystems.com>