Replace the spa_asize_inflation tunable with a calculation
Instead of using a static formula of:
  (VDEV_RAIDZ_MAXPARITY + 1) * SPA_DVAS_PER_BP * 2 == 24

The static formula still accounted for ditto-blocks which have been
removed from ZFS. The formula should have dropped the '* 2' at that
time, resulting in a new value of 12.

Each vdev type now has a function that reports its worse-case
allocation size inflation. For RAID-Z, rather than just multiplying
the logical write size by the minimum RAID-Z allocation size, we
round the logical write size up to the next multiple of the
worst-case minimum allocation size.

Then multiply that by the number of copies for the target dataset.
Falling back to a worse case of SPA_DVAS_PER_BP (3) if the value of
the copies= property cannot be determined.

This replaces the spa_asize_inflation tunable which always assumed
the worst case, and required manual adjustment to get better
performance when near to a quota.

When adding/opening a vdev, set spa->spa_worst_alloc to the worst
case for the worst vdev in the pool.

This will lower the excessive inflation for pools that are simple
stripes or mirror sets.

Replace the other use of spa_asize_inflation (receiving clones with
a refquota) with a static #define SPA_MAX_ASIZE_INFLATION

Sponsored-by: Zededa Inc.
Sponsored-by: Klara Inc.
Signed-off-by: Allan Jude <allan@klarasystems.com>